receive-pack: new external execute-commands hook
Git calls an internal `execute_commands` function to handle commands
sent from client to `receive-pack`.  Regardless of what references
the user pushes, git creates or updates the corresponding references
if write permissions are allowed.  A contributor who has no write
permissions, cannot push to the the repository directly, has to write
commits to an alternate location, and sends pull request by email or
by other ways.  We call this distributed workflow.

It would be more convenient for some cases to work in a centralized
workflow.  Think if there is a filter in front of the internal
`execute_commands` function, a contributor who has no write permission
can send special `git-push` command to let the filter to interpret and
handle.

Add a new hook named "execute-commands" to `receive-pack`, which works
as the filter in front of the internal `execute_commands` funciton.
A contributor may run the following `git-push` command:

    git push -o reviewers=user1,user2 \
        -o oldoid=89c082363ac950d224a7259bfba3ccfbf4c560c4 \
        origin \
        HEAD:refs/for/<branch-name>/<session>

This command has a Gerrit like refspec, which has a target refname with
special prefix ("refs/for/"), and a <session> (usually the local branch
name is used).  Push options and the special refname will pass to the
filter ("execute-commands" hook) first.  Instead of creating or updating
a weird reference, "execute-commands" hook can create a pull request or
send emails for code review.

There are two phases involved to run "execute-commands" hook:

* Run another hook "execute-commands--pre-receive" (or run the
  "execute-command" hook with "--pre-receive" as its parameter in future
  version) to check permissions, just like "pre-receive" hook for
  normal commands for `receive-pack`.

* The "execute-commands" hook will be called (without any parameter), if
  the permission check is ok.  We can use this hook to do some
  interesting work, such as creating pull request for the special
  `git-push` command.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>