refs-hook: pre-check hook to lock repository
Before write to repo, check if repo is locked. User can add a lock file
'agit-repo.lock' in repo dir or any parent directories, and if this file
exists, write to repo is denied.

`git-gc` and `git-pack-refs` are not affected, and will work as usual.

Add test cases in `t/t1491-repo-locked.sh`.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>