pack-redundant: optimize for memory and cpu time
When call pack-redundant witch "--all", if there are too much
local packs and too much redundant objects within them, the
**minimize** step may eat up the memory and cpu because it
would produces a huge-big-size "pll" list in the heap.

Change the strategy of "cmp_two_packs" could solve this: At
first sort all the local packs from smallest to biggest. Then
if two local packs have the same objects, do not remove it from
"unique_objects" in both packs' "pack_list", but just the one
which has bigger size.

With the strategy for "cmp_two_packs" above, the unique set in
the local packs is the collection with smallest size. And
**minimize** will need not to get the permutations for the missing
objects from the unique set, which need less memory and cpu time.

Signed-off-by: sunchao <sunchao9@huawei.com>