test: add option "-p" to create ".git/info"
Commit ead74601c6 (tests: don't assume a .git/info for
.git/info/sparse-checkout, 2022-06-03) introduced a new variable
TEST_CREATE_REPO_NO_TEMPLATE for test framework. New test repository
with this variable would be created from an empty template, and would
create a ".git/info" directory using command:

    mkdir .git/info

It's better to add a option "-p" for it to ignore an already exist
".git/info" directory. This is because the hook like
"reference-transaction" may create ".git/info" directory in advance to
save some meta info in it.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>